Remove SEO.canonical.js as it is no longer required. Replace with appleItunesAppBanner test which only runs on the mundo STY and arabic MAP pages.
