+++ src/integration/pages/articles/canonicalTests.js
@@ -1,7 +1,6 @@
 import {
   runCoreCanonicalTests,
   runCanonicalAnalyticsTests,
-  runSeoCanonicalTests,
 } from '../../common';
 import runCrossPlatformTests from './crossPlatformTests';
 
@@ -9,5 +8,4 @@ export default service => {
   runCrossPlatformTests(service);
   runCoreCanonicalTests();
   runCanonicalAnalyticsTests();
-  runSeoCanonicalTests(service);
 };
