+++ src/integration/pages/idxPage/canonicalTests.js
@@ -2,14 +2,12 @@ import runCrossPlatformTests from './crossPlatformTests';
 import {
   runCoreCanonicalTests,
   runCanonicalAnalyticsTests,
-  runSeoCanonicalTests,
 } from '../../common';
 
 export default service => {
   runCrossPlatformTests(service);
   runCoreCanonicalTests();
   runCanonicalAnalyticsTests();
-  runSeoCanonicalTests(service);
 
   describe('Radio Schedule', () => {
     const hasRadioSchedule = service === 'persian';
