+++ src/integration/pages/liveRadio/canonicalTests.js
@@ -3,7 +3,6 @@ import {
   runCoreCanonicalTests,
   runCanonicalAnalyticsTests,
   runRadioScheduleTests,
-  runSeoCanonicalTests,
 } from '../../common';
 
 export default service => {
@@ -11,5 +10,4 @@ export default service => {
   runCoreCanonicalTests();
   runCanonicalAnalyticsTests();
   runRadioScheduleTests({ isAmp: false });
-  runSeoCanonicalTests(service);
 };
