+++ src/integration/pages/mediaAssetPage/arabicTC2/canonical.test.js
@@ -4,9 +4,11 @@
  */
 
 import runCanonicalTests from '../canonicalTests';
+import { runAppleItunesAppBannerTests } from '../../../common';
 
 describe('Canonical', () => {
   describe(pageType, () => {
     runCanonicalTests(service);
+    runAppleItunesAppBannerTests();
   });
 });
