+++ src/integration/pages/mediaAssetPage/canonicalTests.js
@@ -1,7 +1,6 @@
 import {
   runCoreCanonicalTests,
   runCanonicalAnalyticsTests,
-  runSeoCanonicalTests,
 } from '../../common';
 import runCrossPlatformTests from './crossPlatformTests';
 
@@ -9,7 +8,6 @@ export default service => {
   runCrossPlatformTests(service);
   runCoreCanonicalTests();
   runCanonicalAnalyticsTests();
-  runSeoCanonicalTests(service);
 
   it('I can see the placeholder loading image', () => {
     const loadingImageWrapper = document.querySelector(
