+++ src/integration/pages/onDemandRadioPage/canonicalTests.js
@@ -2,12 +2,10 @@ import {
   runCoreCanonicalTests,
   runCanonicalAnalyticsTests,
   runRadioScheduleTests,
-  runSeoCanonicalTests,
 } from '../../common';
 
 export default () => {
   runCoreCanonicalTests();
   runCanonicalAnalyticsTests();
   runRadioScheduleTests({ isAmp: false });
-  runSeoCanonicalTests(service);
 };
