+++ src/integration/pages/storyPage/canonicalTests.js
@@ -1,7 +1,6 @@
 import {
   runCoreCanonicalTests,
   runCanonicalAnalyticsTests,
-  runSeoCanonicalTests,
 } from '../../common';
 import runCrossPlatformTests from './crossPlatformTests';
 
@@ -9,7 +8,6 @@ export default service => {
   runCrossPlatformTests(service);
   runCoreCanonicalTests();
   runCanonicalAnalyticsTests();
-  runSeoCanonicalTests(service);
 
   describe('Lead image', () => {
     it('I can see an image with a caption', () => {
